+++ UVtools.Core/Operations/OperationPhasedExposure.cs
@@ -121,7 +121,7 @@ public class OperationPhasedExposure : Operation, IEquatable<OperationPhasedExpo
 
     public override string? ValidateSpawn()
     {
-        if (!SlicerFile.CanUseLayerPositionZ || !SlicerFile.CanUseLayerLiftHeight || !SlicerFile.CanUseLayerExposureTime)
+        if (!SlicerFile.CanUseSameLayerPositionZ || !SlicerFile.CanUseLayerLiftHeight || !SlicerFile.CanUseLayerExposureTime)
         {
             return NotSupportedMessage;
         }