+++ UVtools.UI/Controls/RenameFileControl.axaml.cs
@@ -13,6 +13,7 @@ public partial class RenameFileControl : ToolBaseControl
     private string _newFileNameNoExt;
     private readonly string _fileExtension;
     private bool _overwrite;
+    private bool _onlyAsciiCharacters;
 
     public string OldFilePath { get; }
     public string OldFileNameNoExt { get; }
@@ -27,7 +28,7 @@ public partial class RenameFileControl : ToolBaseControl
             if (!RaiseAndSetIfChanged(ref _newFileNameNoExt, value)) return;
             RaisePropertyChanged(nameof(NewFileName));
             RaisePropertyChanged(nameof(NewFilePath));
-            ParentWindow!.ButtonOkEnabled = CanRename;
+            Validate();
         }
     }
 
@@ -35,70 +36,115 @@ public partial class RenameFileControl : ToolBaseControl
 
     public string NewFilePath => Path.Combine(SlicerFile!.DirectoryPath!, NewFileName);
 
+    public bool OnlyAsciiCharacters
+    {
+        get => _onlyAsciiCharacters;
+        set
+        {
+            if(!RaiseAndSetIfChanged(ref _onlyAsciiCharacters, value)) return;
+            Validate();
+        }
+    }
+
     public bool Overwrite
     {
         get => _overwrite;
         set => RaiseAndSetIfChanged(ref _overwrite, value);
     }
 
-    public bool CanRename => !string.IsNullOrWhiteSpace(_newFileNameNoExt) && !string.Equals(OldFileNameNoExt, _newFileNameNoExt.Trim(), StringComparison.Ordinal) && _newFileNameNoExt.IndexOfAny(Path.GetInvalidFileNameChars()) == -1;
-
     public RenameFileControl()
     {
         OldFilePath = SlicerFile!.FileFullPath!;
         _newFileNameNoExt = OldFileNameNoExt = FileFormat.GetFileNameStripExtensions(OldFilePath, out _fileExtension);
+        _onlyAsciiCharacters = UserSettings.Instance.Automations.FileNameOnlyAsciiCharacters;
             
         DataContext = this;
         InitializeComponent();
     }
 
-    protected override void OnInitialized()
+    public RenameFileControl(string newFileNameNoExt) : this()
     {
-        ParentWindow!.ButtonOkEnabled = false;
+        _newFileNameNoExt = newFileNameNoExt;
     }
 
-    public override string Validate()
+    protected override void OnInitialized()
+    {
+        ParentWindow!.IsROIOrMasksVisible = false;
+        Validate();
+    }
+
+    protected override string? ValidateInternally()
     {
         var sb = new StringBuilder();
 
-        if (string.IsNullOrWhiteSpace(_newFileNameNoExt))
+        var newFileName = _newFileNameNoExt.Trim();
+        if (string.IsNullOrWhiteSpace(newFileName))
         {
-            sb.AppendLine("The filename can not be blank.");
+            return "The file name can not be blank.";
         }
 
-        if (string.Equals(OldFileNameNoExt, _newFileNameNoExt, StringComparison.Ordinal))
+        if (string.Equals(OldFileNameNoExt, newFileName, StringComparison.Ordinal))
         {
-            sb.AppendLine("The old and new filename can not be the same.");
+            return "The file name must be different from the previous.";
         }
 
-        var invalidChars = Path.GetInvalidFileNameChars();
-        if (_newFileNameNoExt.IndexOfAny(invalidChars) >= 0)
+        if (!FileFormat.IsFileNameValid(newFileName, out var message, _onlyAsciiCharacters))
         {
-            return $"The filename have invalid characters. The following characters are forbidden:\n{string.Join(", ", Path.GetInvalidFileNameChars())}";
+            return message;
         }
+
+        /*var invalidChars = Path.GetInvalidFileNameChars();
+        if (newFileName.IndexOfAny(invalidChars) >= 0)
+        {
+            return $"The filename have invalid characters. The following characters are forbidden:\n{string.Join(", ", invalidChars)}";
+        }
+
+        if (_onlyAsciiCharacters)
+        {
+            var nonAscii = newFileName.Where(c => !char.IsAscii(c)).Distinct();
+            if (nonAscii.Any())
+            {
+                return $"The filename have non-ASCII characters. The following characters are forbidden:\n{string.Join(", ", nonAscii)}";
+            }
+        }*/
 
         return sb.ToString();
     }
 
-    public override async Task<bool> OnBeforeProcess()
+    protected override void OnAfterValidate()
+    {
+        ParentWindow!.ButtonOkEnabled = IsLastValidationSuccess;
+    }
+
+    public override async Task<bool> OnAfterProcess()
     {
         NewFileNameNoExt = _newFileNameNoExt.Trim();
-        if (!CanRename)
-        {
-            return false;
-        }
+        if (!IsLastValidationSuccess) return false;
 
-        if (_overwrite) return true;
+        if (_overwrite) return await RenameCurrentSlicerFile();
 
-        if (!File.Exists(NewFilePath)) return true;
+        if (!File.Exists(NewFilePath)) return await RenameCurrentSlicerFile();
         if (await ParentWindow!.MessageBoxQuestion(
                 $"The file \"{_newFileNameNoExt}\" already exists, do you want to overwrite?",
                 "File already exists") == MessageButtonResult.Yes)
         {
             Overwrite = true;
-            return true;
+            return await RenameCurrentSlicerFile();
         }
 
         return false;
+    }
+
+    private async Task<bool> RenameCurrentSlicerFile()
+    {
+        try
+        {
+            return SlicerFile!.RenameFile(_newFileNameNoExt, _overwrite);
+        }
+        catch (Exception e)
+        {
+            await ParentWindow!.MessageBoxError(e.ToString(), "Error while trying to rename the file");
+            return false;
+        }
     }
 }